+++ src/javascript/SelectorTypes/SelectorTypes.js
@@ -6,7 +6,8 @@ import ChoiceList from './ChoiceList';
 import registerChoiceListActions from './ChoiceList/ChoiceList.actions';
 import registerChoiceListOnChange from './ChoiceList/ChoiceList.onChange';
 import DateTimePicker from './DateTimePicker';
-import {getPickerSelectorType, registerPickerConfig} from './Picker/Picker.configs';
+import {registerPickerConfig} from './Picker/Picker.configs';
+import {getPickerSelectorType} from './Picker/Picker.utils';
 import Checkbox from './Checkbox';
 import Category from './Category';
 import {registerPickerActions} from './Picker/actions';
