+++ src/javascript/SelectorTypes/SelectorTypes.spec.js
@@ -54,11 +54,11 @@ describe('Selector Types', () => {
                 selectorOptions: [{name: 'type', value: 'editorial'}]
             });
             expect(selector.cmp).toEqual(Picker);
-            expect(selector.key).toEqual('ContentPicker');
+            expect(selector.key).toEqual('ContentPickerSelectorType');
 
             selector = resolveSelectorType({selectorType: 'Picker', selectorOptions: [{name: 'type', value: 'image'}]});
             expect(selector.cmp).toEqual(Picker);
-            expect(selector.key).toEqual('MediaPicker');
+            expect(selector.key).toEqual('MediaPickerSelectorType');
         });
 
         it('should adapt value if the selector option is password', () => {
